fix: 🐛 incorrect cookie refresh
